refactor cancel tests out to specific test class
