+++ src/Microsoft.DotNet.Interactive.Tests/CancelCommandTests.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System;
+using System.Linq;
 using System.Threading.Tasks;
 using FluentAssertions;
 using FluentAssertions.Execution;
@@ -54,12 +56,15 @@ namespace Microsoft.DotNet.Interactive.Tests
                 .Be(cancelCommand);
         }
 
-        [Fact]
-        public async Task cancel_command_cancels_all_deferred_commands_on_subkernels()
+        [Theory]
+        [InlineData(Language.CSharp)]
+        [InlineData(Language.FSharp)]
+        [InlineData(Language.PowerShell)]
+        public async Task cancel_command_cancels_all_deferred_commands_on_subkernels(Language language)
         {
             var deferredCommandExecuted = false;
 
-            var kernel = CreateKernel();
+            var kernel = CreateKernel(language);
 
             var deferred = new SubmitCode("placeholder")
             {
@@ -72,7 +77,10 @@ namespace Microsoft.DotNet.Interactive.Tests
 
             var cancelCommand = new Cancel();
 
-            kernel.ChildKernels[0].DeferCommand(deferred);
+            foreach (var subkernel in kernel.ChildKernels)
+            {
+                subkernel.DeferCommand(deferred);
+            }
 
             await kernel.SendAsync(cancelCommand);
 
@@ -88,5 +96,87 @@ namespace Microsoft.DotNet.Interactive.Tests
                 .Be(cancelCommand);
         }
 
+        [Theory]
+        [InlineData(Language.CSharp)]
+        [InlineData(Language.FSharp)]
+        [InlineData(Language.PowerShell, Skip = "to address later")]
+        public void cancel_command_cancels_the_running_command(Language language)
+        {
+            var submitCodeCommandCancelled = false;
+            var submitCodeCommandStarted = false;
+            var kernel = CreateKernel(language);
+
+            var cancelCommand = new Cancel();
+
+            var submitCodeCommand = new SubmitCode("placeholder")
+            {
+                Handler = (command, context) =>
+                {
+                    submitCodeCommandStarted = true;
+
+                    while (!context.CancellationToken.IsCancellationRequested)
+                    {
+
+                    }
+
+                    submitCodeCommandCancelled = true;
+
+                    return Task.CompletedTask;
+                }
+            };
+
+            Task.WhenAll(
+                    Task.Run(async () =>
+                    {
+                        await kernel.SendAsync(submitCodeCommand);
+                    }),
+                    Task.Run(async () =>
+                    {
+                        await Task.Delay(100);
+                        await kernel.SendAsync(cancelCommand);
+                    })).Wait();
+            //  .Wait(TimeSpan.FromSeconds(20));
+
+
+            using var _ = new AssertionScope();
+
+            submitCodeCommandStarted.Should().BeTrue();
+            submitCodeCommandCancelled.Should().BeTrue();
+
+            KernelEvents
+                .Should()
+                .ContainSingle<CommandFailed>(c => c.Command == submitCodeCommand)
+                .Which
+                .Exception
+                .Should()
+                .BeOfType<OperationCanceledException>();
+        }
+
+
+        [Theory]
+        [InlineData(Language.CSharp, "await Task.Delay(Microsoft.DotNet.Interactive.KernelInvocationContext.Current.Token); Console.WriteLine(\"done c#\")", "done c#")]
+        [InlineData(Language.FSharp, @"
+System.Threading.Thread.Sleep(3000)
+Console.WriteLine(""done c#"")", "done f#")]
+        public async Task user_code_can_react_to_cancel_command_using_cancellation_token(Language language, string code, string expectedValue)
+        {
+            var kernel = CreateKernel(language);
+            var cancelCommand = new Cancel();
+            var submitCodeCommand = new SubmitCode(code);
+
+            Task.Run(async () => await kernel.SendAsync(submitCodeCommand));
+            await Task.Delay(100);
+            await kernel.SendAsync(cancelCommand);
+
+
+            KernelEvents
+                .Should()
+                .ContainSingle<StandardOutputValueProduced>()
+                .Which
+                .Value
+                .Should()
+                .Be(expectedValue);
+
+        }
     }
 }
