+++ src/Microsoft.DotNet.Interactive.Tests/LanguageKernelTests.cs
@@ -1267,209 +1267,5 @@ Console.Write(2);
             succeeded.Should().BeTrue();
             x.Should().Be("hello");
         }
-
-
-        [Theory]
-        [InlineData(Language.CSharp, "System.Threading.Thread.Sleep(3000);")]
-        [InlineData(Language.FSharp, "System.Threading.Thread.Sleep(3000)")]
-        [InlineData(Language.PowerShell, "Start-Sleep -Milliseconds 3000", Skip = "to address later")]
-        public void Quit_command_is_handled(Language language, string code)
-        {
-            var kernel = CreateKernel(language);
-
-            var quitCommandExecuted = false;
-            
-            Quit.OnQuit(() => { quitCommandExecuted = true; });
-
-            var quitCommand = new Quit();
-
-            var submitCodeCommand = new SubmitCode(code);
-
-            Task.WhenAll(
-                    Task.Run(async () =>
-                    {
-                        await Task.Delay(20);
-                        await kernel.SendAsync(submitCodeCommand);
-                    }),
-                    Task.Run(async () =>
-                    {
-                        await Task.Delay(100);
-                        await kernel.SendAsync(quitCommand);
-                    }))
-                .Wait(TimeSpan.FromSeconds(20));
-
-            using var _ = new AssertionScope();
-
-            quitCommandExecuted.Should().BeTrue();
-
-            KernelEvents
-                .Should()
-                .ContainSingle<CommandSucceeded>()
-                .Which
-                .Command
-                .Should()
-                .Be(quitCommand);
-        }
-
-        [Theory]
-        [InlineData(Language.CSharp, "System.Threading.Thread.Sleep(3000);")]
-        [InlineData(Language.FSharp, "System.Threading.Thread.Sleep(3000)")]
-        [InlineData(Language.PowerShell, "Start-Sleep -Milliseconds 3000", Skip = "to address later")]
-        public void Quit_command_causes_the_running_command_to_fail(Language language, string code)
-        {
-            var kernel = CreateKernel(language);
-
-            Quit.OnQuit(() => { });
-
-            var quitCommand = new Quit();
-
-            var submitCodeCommand = new SubmitCode(code);
-
-            Task.WhenAll(
-                Task.Run(async () =>
-                {
-                    await Task.Delay(20);
-                    await kernel.SendAsync(submitCodeCommand);
-                }),
-                Task.Run(async () =>
-                {
-                    await Task.Delay(100);
-                    await kernel.SendAsync(quitCommand);
-                }))
-                .Wait(TimeSpan.FromSeconds(20));
-
-            using var _ = new AssertionScope();
-
-            KernelEvents
-                .Should()
-                .ContainSingle<CommandFailed>(c => c.Command == submitCodeCommand)
-                .Which
-                .Exception
-                .Should()
-                .BeOfType<OperationCanceledException>();
-        }
-
-
-        [Theory]
-        [InlineData(Language.CSharp, "await Task.Delay(Microsoft.DotNet.Interactive.KernelInvocationContext.Current.Token); Console.WriteLine(\"done c#\")", "done c#")]
-        [InlineData(Language.FSharp, @"
-System.Threading.Thread.Sleep(3000)
-Console.WriteLine(""done c#"")", "done f#")]
-        public async Task user_code_can_react_to_cancel_command_using_cancellation_token(Language language, string code, string expectedValue)
-        {
-            var kernel = CreateKernel(language);
-            var cancelCommand = new Cancel();
-            var submitCodeCommand = new SubmitCode(code);
-
-            Task.Run(async () => await kernel.SendAsync(submitCodeCommand));
-            await Task.Delay(100);
-            await kernel.SendAsync(cancelCommand);
-
-
-            KernelEvents
-                .Should()
-                .ContainSingle<StandardOutputValueProduced>()
-                .Which
-                .Value
-                .Should()
-                .Be(expectedValue);
-
-        }
-
-        [Theory]
-        [InlineData(Language.CSharp)]
-        [InlineData(Language.FSharp)]
-        [InlineData(Language.PowerShell, Skip = "to address later")]
-        public void cancel_command_cancels_the_running_command(Language language)
-        {
-            var submitCodeCommandCancelled = false;
-            var submitCodeCommandStarted = false;
-            var kernel = CreateKernel(language);
-
-            var cancelCommand = new Cancel();
-
-            var submitCodeCommand = new SubmitCode("placeholder")
-            {
-                Handler = (command, context) =>
-                {
-                    submitCodeCommandStarted = true;
-
-                    while (!context.CancellationToken.IsCancellationRequested)
-                    {
-
-                    }
-
-                    submitCodeCommandCancelled = true;
-
-                    return Task.CompletedTask;
-                }
-            };
-
-            Task.WhenAll(
-                    Task.Run(async () =>
-                    {
-                        await kernel.SendAsync(submitCodeCommand);
-                    }),
-                    Task.Run(async () =>
-                    {
-                        await Task.Delay(100);
-                        await kernel.SendAsync(cancelCommand);
-                    })).Wait();
-              //  .Wait(TimeSpan.FromSeconds(20));
-
-
-            using var _ = new AssertionScope();
-
-            submitCodeCommandStarted.Should().BeTrue();
-            submitCodeCommandCancelled.Should().BeTrue();
-
-            KernelEvents
-                .Should()
-                .ContainSingle<CommandFailed>(c => c.Command == submitCodeCommand)
-                .Which
-                .Exception
-                .Should()
-                .BeOfType<OperationCanceledException>();
-        }
-
-        [Theory]
-        [InlineData(Language.CSharp)]
-        [InlineData(Language.FSharp)]
-        public async Task cancel_command_cancels_all_deferred_commands(Language language)
-        {
-            var deferredCommandExecuted = false;
-
-            var kernel = CreateKernel(language);
-            var languageKernel = kernel.ChildKernels.OfType<DotNetKernel>().Single();
-
-
-            var deferred = new SubmitCode("placeholder")
-            {
-                Handler = (command, context) =>
-                {
-                    deferredCommandExecuted = true;
-                    return Task.CompletedTask;
-                }
-            };
-
-            var cancelCommand = new Cancel();
-
-            languageKernel.DeferCommand(deferred);
-
-            var events = kernel.KernelEvents.ToSubscribedList();
-
-            await kernel.SendAsync(cancelCommand);
-
-            using var _ = new AssertionScope();
-
-            deferredCommandExecuted.Should().BeFalse();
-
-            events
-                .Should().ContainSingle<CommandSucceeded>()
-                .Which
-                .Command
-                .Should()
-                .Be(cancelCommand);
-        }
     }
 }
