+++ src/Microsoft.DotNet.Interactive.Tests/QuitCommandTests.cs
@@ -87,14 +87,17 @@ namespace Microsoft.DotNet.Interactive.Tests
                 .Be(quitCommand);
         }
 
-        [Fact]
-        public async Task quit_command_cancels_all_deferred_commands_on_subkernels()
+        [Theory]
+        [InlineData(Language.CSharp)]
+        [InlineData(Language.FSharp)]
+        [InlineData(Language.PowerShell)]
+        public async Task quit_command_cancels_all_deferred_commands_on_subkernels(Language language)
         {
             var deferredCommandExecuted = false;
 
             var quitCommandExecuted = false;
 
-            var kernel = CreateKernel();
+            var kernel = CreateKernel(language);
 
             var deferred = new SubmitCode("placeholder")
             {
@@ -109,7 +112,10 @@ namespace Microsoft.DotNet.Interactive.Tests
 
             var quitCommand = new Quit();
 
-            kernel.ChildKernels[0].DeferCommand(deferred);
+            foreach (var subkernel in kernel.ChildKernels)
+            {
+                subkernel.DeferCommand(deferred);
+            }
 
             await kernel.SendAsync(quitCommand);
 
@@ -124,6 +130,87 @@ namespace Microsoft.DotNet.Interactive.Tests
                 .Command
                 .Should()
                 .Be(quitCommand);
+        }
+
+        [Theory]
+        [InlineData(Language.CSharp, "System.Threading.Thread.Sleep(3000);")]
+        [InlineData(Language.FSharp, "System.Threading.Thread.Sleep(3000)")]
+        [InlineData(Language.PowerShell, "Start-Sleep -Milliseconds 3000", Skip = "to address later")]
+        public void Quit_command_is_handled(Language language, string code)
+        {
+            var kernel = CreateKernel(language);
+
+            var quitCommandExecuted = false;
+
+            Quit.OnQuit(() => { quitCommandExecuted = true; });
+
+            var quitCommand = new Quit();
+
+            var submitCodeCommand = new SubmitCode(code);
+
+            Task.WhenAll(
+                    Task.Run(async () =>
+                    {
+                        await Task.Delay(20);
+                        await kernel.SendAsync(submitCodeCommand);
+                    }),
+                    Task.Run(async () =>
+                    {
+                        await Task.Delay(100);
+                        await kernel.SendAsync(quitCommand);
+                    }))
+                .Wait(TimeSpan.FromSeconds(20));
+
+            using var _ = new AssertionScope();
+
+            quitCommandExecuted.Should().BeTrue();
+
+            KernelEvents
+                .Should()
+                .ContainSingle<CommandSucceeded>()
+                .Which
+                .Command
+                .Should()
+                .Be(quitCommand);
+        }
+
+
+        [Theory]
+        [InlineData(Language.CSharp, "System.Threading.Thread.Sleep(3000);")]
+        [InlineData(Language.FSharp, "System.Threading.Thread.Sleep(3000)")]
+        [InlineData(Language.PowerShell, "Start-Sleep -Milliseconds 3000", Skip = "to address later")]
+        public void Quit_command_causes_the_running_command_to_fail(Language language, string code)
+        {
+            var kernel = CreateKernel(language);
+
+            Quit.OnQuit(() => { });
+
+            var quitCommand = new Quit();
+
+            var submitCodeCommand = new SubmitCode(code);
+
+            Task.WhenAll(
+                    Task.Run(async () =>
+                    {
+                        await Task.Delay(20);
+                        await kernel.SendAsync(submitCodeCommand);
+                    }),
+                    Task.Run(async () =>
+                    {
+                        await Task.Delay(100);
+                        await kernel.SendAsync(quitCommand);
+                    }))
+                .Wait(TimeSpan.FromSeconds(20));
+
+            using var _ = new AssertionScope();
+
+            KernelEvents
+                .Should()
+                .ContainSingle<CommandFailed>(c => c.Command == submitCodeCommand)
+                .Which
+                .Exception
+                .Should()
+                .BeOfType<OperationCanceledException>();
         }
     }
 }
